+++ src/connect.rs
@@ -1214,11 +1214,12 @@ mod rustls_tls_conn {
 mod socks {
     use std::io;
 
+    use http::uri::Scheme;
     use http::Uri;
     use tokio::net::TcpStream;
     use tokio_socks::tcp::{Socks4Stream, Socks5Stream};
 
-    use super::{BoxError, Scheme};
+    use super::BoxError;
     use crate::proxy::Intercepted;
 
     pub(super) enum DnsResolve {
@@ -1288,6 +1289,7 @@ mod socks {
 }
 
 mod verbose {
+    use crate::util::Escape;
     use hyper::rt::{Read, ReadBufCursor, Write};
     use hyper_util::client::legacy::connect::{Connected, Connection};
     use std::cmp::min;
@@ -1338,7 +1340,7 @@ mod verbose {
             let mut vbuf = hyper::rt::ReadBuf::uninit(unsafe { buf.as_mut() });
             match Pin::new(&mut self.inner).poll_read(cx, vbuf.unfilled()) {
                 Poll::Ready(Ok(())) => {
-                    log::trace!("{:08x} read: {:?}", self.id, Escape(vbuf.filled()));
+                    log::trace!("{:08x} read: {:?}", self.id, Escape::new(vbuf.filled()));
                     let len = vbuf.filled().len();
                     // SAFETY: The two cursors were for the same buffer. What was
                     // filled in one is safe in the other.
@@ -1361,7 +1363,7 @@ mod verbose {
         ) -> Poll<Result<usize, std::io::Error>> {
             match Pin::new(&mut self.inner).poll_write(cx, buf) {
                 Poll::Ready(Ok(n)) => {
-                    log::trace!("{:08x} write: {:?}", self.id, Escape(&buf[..n]));
+                    log::trace!("{:08x} write: {:?}", self.id, Escape::new(&buf[..n]));
                     Poll::Ready(Ok(n))
                 }
                 Poll::Ready(Err(e)) => Poll::Ready(Err(e)),
@@ -1414,35 +1416,6 @@ mod verbose {
         }
     }
 
-    struct Escape<'a>(&'a [u8]);
-
-    impl fmt::Debug for Escape<'_> {
-        fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-            write!(f, "b\"")?;
-            for &c in self.0 {
-                // https://doc.rust-lang.org/reference.html#byte-escapes
-                if c == b'\n' {
-                    write!(f, "\\n")?;
-                } else if c == b'\r' {
-                    write!(f, "\\r")?;
-                } else if c == b'\t' {
-                    write!(f, "\\t")?;
-                } else if c == b'\\' || c == b'"' {
-                    write!(f, "\\{}", c as char)?;
-                } else if c == b'\0' {
-                    write!(f, "\\0")?;
-                // ASCII printable
-                } else if c >= 0x20 && c < 0x7f {
-                    write!(f, "{}", c as char)?;
-                } else {
-                    write!(f, "\\x{c:02x}")?;
-                }
-            }
-            write!(f, "\"")?;
-            Ok(())
-        }
-    }
-
     struct Vectored<'a, 'b> {
         bufs: &'a [IoSlice<'b>],
         nwritten: usize,
@@ -1456,7 +1429,7 @@ mod verbose {
                     break;
                 }
                 let n = min(left, buf.len());
-                Escape(&buf[..n]).fmt(f)?;
+                Escape::new(&buf[..n]).fmt(f)?;
                 left -= n;
             }
             Ok(())